Add files via upload
main objective is to solve the challenge using OOPS concept

4 balls move from left to right

ball 1 is at 1/5th height from the top, with the 1 units per frame. ball 2 is at 2/5th height from the top, with the 2 units per frame. ball 3 is at 3/5th height from the top, with the 3 units per frame. ball 4 is at 4/5th height from the top, with the 4 units per frame.